pc - dry up test of whether bootstrap is loaded
