+++ pom.xml
@@ -3,8 +3,8 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.springframework</groupId>
-    <artifactId>gs-serving-web-content</artifactId>
+    <groupId>edu.ucsb.cs56</groupId>
+    <artifactId>ucsb_courses_search</artifactId>
     <version>0.1.0</version>
 
     <parent>
