+++ src/test/java/edu/ucsb/cs56/ucsb_courses_search/GreetingControllerTest.java
@@ -1,4 +1,4 @@
-package hello;
+package edu.ucsb.cs56.ucsb_courses_search;
 
 // import static org.hamcrest.Matchers.equalTo;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
@@ -19,48 +19,38 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @AutoConfigureMockMvc
-public class HomePageTest {
+public class GreetingControllerTest {
 
     @Autowired
     private MockMvc mvc;
 
     @Test
-    public void getHomePage_ContentType() throws Exception {
-        mvc.perform(MockMvcRequestBuilders.get("/").accept(MediaType.TEXT_HTML))
+    public void getGreeting_ContentType() throws Exception {
+        mvc.perform(MockMvcRequestBuilders.get("/greeting").accept(MediaType.TEXT_HTML))
                 .andExpect(status().isOk())
                 .andExpect(content().contentType("text/html;charset=UTF-8"));
     }
 
-
     @Test
-    public void getHomePage_BootstrapLoaded() throws Exception {
-        mvc.perform(MockMvcRequestBuilders.get("/").accept(MediaType.TEXT_HTML))
-                .andExpect(status().isOk())
-                .andExpect(xpath(BootstrapLiterals.bootstrapCSSXpath).exists());
-        for (String s: BootstrapLiterals.bootstrapJSurls) {
-            String jsXPath = String.format("//script[@src='%s']",s);
-            mvc.perform(MockMvcRequestBuilders.get("/greeting").accept(MediaType.TEXT_HTML))
-              .andExpect(status().isOk())
-              .andExpect(xpath(jsXPath).exists());
-        }
+    public void getGreeting_BootstrapLoaded() throws Exception {
+       BootstrapTestHelper.bootstrapIsLoaded(mvc, "/greeting");
     }
 
-
     @Test
-    public void getHomePage_hasCorrectTitle() throws Exception {
-        mvc.perform(MockMvcRequestBuilders.get("/").accept(MediaType.TEXT_HTML))
+    public void getGreeting_hasCorrectTitle() throws Exception {
+        mvc.perform(MockMvcRequestBuilders.get("/greeting").accept(MediaType.TEXT_HTML))
                 .andExpect(status().isOk())
                 .andExpect(xpath("//title").exists())
                 .andExpect(xpath("//title").string("Getting Started: Serving Web Content"));
     }
 
     @Test
-    public void getHomePage_hasNavBar() throws Exception {
-        NavigationTestHelper.hasNavBar(mvc, "/");
+    public void getGreeting_hasNavBar() throws Exception {
+        NavigationTestHelper.hasNavBar(mvc, "/greeting");
     }
 
     @Test
-    public void getHomePage_hasFooter() throws Exception {
-        NavigationTestHelper.hasFooter(mvc, "/");
+    public void getGreeting_hasFooter() throws Exception {
+        NavigationTestHelper.hasFooter(mvc, "/greeting");
     }
 }
